# haven/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Haven API
  version: 0.1.0
  description: |
    Haven is an AI-powered housing marketplace matching seekers and landlords
    based on lifestyle compatibility, verified trust scores, and real-time chat.
  contact:
    name: Haven Engineering
    email: engineering@haven.app
servers:
  - url: https://haven.app
    description: Production
  - url: https://staging.haven.app
    description: Staging
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        user_type:
          type: string
          enum: [seeker, landlord, both]
        onboarding_completed:
          type: boolean
...
paths:
...
  /api/users/profile:
    get:
      operationId: getProfile
      summary: Get current user profile
      tags: [Users]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    patch:
      operationId: updateProfile
      summary: Update current user profile
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Updated profile
...
  /api/users/gdpr:
    get:
      operationId: exportUserData
      summary: "GDPR Article 15: Export all personal data"
      tags: [Users]
      responses:
        '200':
          description: JSON file containing all personal data
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized

    delete:
      operationId: deleteUserData
      summary: "GDPR Article 17: Erase all personal data"
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: string
                  enum: ["DELETE MY ACCOUNT"]
                  description: Must be the exact string "DELETE MY ACCOUNT"
                hard_delete:
                  type: boolean
                  default: false
                  description: If true, permanently deletes the auth account. If false, anonymises data.
                preserve_audit_logs:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Data deleted or anonymised
        '400':
          description: Confirm string missing or wrong
        '401':
          description: Unauthorized
...